fix(events): use u32 instead of usize for count fields
Change usize fields to u32 in event data types for protobuf Struct
compatibility. When numbers pass through protobuf Struct (which uses
double), they may serialize as floats (0.0) which fails to
deserialize as usize. u32 handles this correctly.

Affected fields:
- ReasonCompletedData.tool_call_count
- ActCompletedData.success_count, error_count
- ActResult.success_count, error_count
- TurnCompletedData.iterations

+++ crates/core/src/atoms/act.rs
@@ -69,9 +69,9 @@ pub struct ActResult {
     /// Whether all tool calls completed (regardless of success/failure)
     pub completed: bool,
     /// Number of successful tool calls
-    pub success_count: usize,
+    pub success_count: u32,
     /// Number of failed tool calls
-    pub error_count: usize,
+    pub error_count: u32,
 }
 
 // ============================================================================
@@ -208,8 +208,8 @@ where
         let results = join_all(futures).await;
 
         // Count successes and errors
-        let success_count = results.iter().filter(|r| r.success).count();
-        let error_count = results.iter().filter(|r| !r.success).count();
+        let success_count = results.iter().filter(|r| r.success).count() as u32;
+        let error_count = results.iter().filter(|r| !r.success).count() as u32;
 
         // Emit act.completed event
         if let Err(e) = self
